# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "",
    "description": "",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8000",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "tempTokenHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "tempToken",
      "description": "Temporary token for two-factor authentication"
    },
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    },
    "refreshTokenHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "refreshToken",
      "description": "Refresh token for authentication"
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "register",
        "tags": ["/auth"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "John" },
                "surname": { "type": "string", "example": "Doe" },
                "phone": { "type": "string", "example": "+905079026718" },
                "role": {
                  "type": "string",
                  "enum": ["admin", "driver", "customer"],
                  "example": "admin"
                },
                "password": { "type": "string", "example": "adminadmin" }
              },
              "required": ["name", "surname", "phone", "role", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User registered successfully"
                }
              }
            }
          },
          "401": {
            "description": "User already exists",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User already exists"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create user",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to create user"
                }
              }
            }
          }
        }
      }
    },
    "/auth/two-factor-auth": {
      "post": {
        "operationId": "twoFactorAuth",
        "tags": ["/auth"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phone": { "type": "string", "example": "+905079026718" },
                "password": { "type": "string", "example": "adminadmin" }
              },
              "required": ["phone", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Two factor authentication successful",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "surname": { "type": "string" },
                    "phone": { "type": "string" },
                    "role": {
                      "type": "string",
                      "enum": ["admin", "driver", "customer"]
                    }
                  }
                },
                "tempToken": { "type": "string" },
                "message": { "type": "string" }
              }
            },
            "examples": {
              "application/json": {
                "user": {
                  "name": "John",
                  "surname": "Doe",
                  "phone": "+905079026718",
                  "role": "admin"
                },
                "tempToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                "message": "Two factor code sent successfully"
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid credentials"
                }
              }
            }
          },
          "500": {
            "description": "Failed to send two factor code",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to send two factor code"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-two-factor-code": {
      "post": {
        "operationId": "verifyTwoFactorCode",
        "tags": ["/auth"],
        "security": [
          {
            "tempTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "string", "example": "123456" }
              },
              "required": ["code"]
            }
          }
        ],

        "responses": {
          "200": {
            "description": "Two factor authentication successful",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string", "example": "John" },
                    "surname": { "type": "string", "example": "Doe" },
                    "phone": { "type": "string", "example": "+905079026718" },
                    "role": {
                      "type": "string",
                      "enum": ["admin", "driver", "customer"],
                      "example": "admin"
                    }
                  }
                },
                "accessToken": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "refreshToken": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expiresIn": { "type": "number", "example": 86400 },
                "refreshTokenExpiresIn": { "type": "number", "example": 604800 }
              },
              "required": ["user", "accessToken", "refreshToken", "expiresIn"]
            }
          },
          "401": {
            "description": "Invalid code",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Invalid code" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/send-reset-password-link": {
      "post": {
        "operationId": "sendResetPasswordLink",
        "tags": ["/auth"],
        "security": [
          {
            "tempTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phone": { "type": "string", "example": "+905079026718" }
              },
              "required": ["phone"]
            }
          }
        ],

        "responses": {
          "200": {
            "description": "Reset password link sent successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Reset password link sent successfully"
                },
                "expiresAt": {
                  "type": "string",
                  "example": "2025-08-11T12:00:00.000Z"
                }
              },
              "required": ["message", "expiresAt"]
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid credentials"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to create reset password link"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "tags": ["/auth"],
        "security": [
          {
            "tempTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phone": { "type": "string", "example": "+905079026718" },
                "token": {
                  "type": "string",
                  "example": "a658ea02f5e6d05ca36a2be90bb7e0..."
                },
                "newPassword": {
                  "type": "string",
                  "example": "adminadmin"
                }
              },
              "required": ["phone", "token", "newPassword"]
            }
          }
        ],

        "responses": {
          "200": {
            "description": "Password updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Password updated successfully"
                }
              },
              "required": ["message"]
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid credentials"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to update password"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh-token": {
      "get": {
        "operationId": "refreshToken",
        "tags": ["/auth"],
        "security": [
          {
            "refreshTokenHeader": [],
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Two factor authentication successful",
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "refreshToken": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expiresIn": { "type": "number", "example": 86400 },
                "refreshTokenExpiresIn": { "type": "number", "example": 604800 }
              },
              "required": [
                "accessToken",
                "refreshToken",
                "expiresIn",
                "refreshTokenExpiresIn"
              ]
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid credentials"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to update password"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "operationId": "logout",
        "tags": ["/auth"],
        "security": [
          {
            "refreshTokenHeader": [],
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Logged out successfully"
                }
              },
              "required": ["message"]
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid refresh token"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to revoke refresh token"
                }
              }
            }
          }
        }
      }
    }
  }
}